+++ sync-for-expenses/CodatSyncExpenses/Models/Shared/Transaction.cs
@@ -31,7 +31,7 @@ namespace Codat.Sync.Expenses.Models.Shared
         /// Status of the transaction.
         /// </summary>
         [JsonProperty("status")]
-        public TransactionStatus? Status { get; set; }
+        public Models.Shared.TransactionStatus? Status { get; set; }
 
         /// <summary>
         /// Your unique idenfier of the transaction.